+++ api/src/test/java/org/openmrs/module/radiology/report/RadiologyReportValidatorTest.java
@@ -7,18 +7,18 @@
  * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
  * graphic logo is a trademark of OpenMRS Inc.
  */
-package org.openmrs.module.radiology.validator;
+package org.openmrs.module.radiology.report;
 
 import static org.hamcrest.Matchers.is;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertThat;
 import static org.junit.Assert.assertTrue;
+
 import org.junit.Test;
 import org.openmrs.Provider;
 import org.openmrs.module.radiology.PerformedProcedureStepStatus;
 import org.openmrs.module.radiology.RadiologyOrder;
 import org.openmrs.module.radiology.Study;
-import org.openmrs.module.radiology.report.RadiologyReport;
 import org.springframework.validation.BindException;
 import org.springframework.validation.Errors;
